Tidy up further
